Issue #221 remove -a flag in install
Specify the -a flag is not the expected behavior for install. This
forces a recompile of all code, even in the standard library. This does
not work when you're doing static compiles using CGO_ENABLED=0 on linux
because typically the end user does not own those files.

This change removes the '-a' flag as this is only really supposed to be
used by the developers of go.

@see https://github.com/golang/go/issues/15450
@see https://golang.org/cl/10761