Do not run non-matching filtered commands.
This changes the behavior of the execution of Go commands. If a command
is run with a filter, and the filter does not match any items, rather
than running the command without specifying any packages (usually
counterintuitively causing Go to assume the current package should be
used), we simply exit without an error.