Update README.md to reflect that ignored packages cannot end in "/"
At https://github.com/kardianos/govendor/blob/e07957427183a9892f35634ffc9ea48dedc6bbb4/context/resolve.go#L215, we add "/" to the prefix expected for excluded package names, so package ignores ending with "/" will not match.  I'm not sure if this was intended or not -- perhaps we should be stripping the trailing slash off before we store the excluded package name -- but it is current behavior.